Skipping Dialogue bug fix
This bug was produced becauese nextPressed in DialogueUI would be set to true, but would not reset to false. On the second run of the dialogue, since nextPressed was true, the RunNode in DialogueManager would fully execute, ending the Dialogue on a single node dialogue. Additionally, removed any visible text from the Dialogue prefab to ensure no text would be displayed when the dialogue is deferred for a single frame.